change processor's id to 64bit int, to avoid assert failed

+++ Processor/include/BoostProcessor.h
@@ -36,7 +36,7 @@ namespace Processor
         void stop();
 
 		struct event* addLocalTimer(
-                const unsigned theId, 
+                const unsigned long long theId, 
 				const struct timeval& theInterval, 
 				event_callback_fn theCallback,
 				void* theArg)
@@ -45,7 +45,7 @@ namespace Processor
             return workersM[workerId].addLocalTimer(theInterval, theCallback, theArg);
         }
 		inline void cancelLocalTimer(
-                const unsigned theId, 
+                const unsigned long long theId, 
                 struct event*& theEvent)
         {
             unsigned workerId = theId % threadCountM;
@@ -53,30 +53,30 @@ namespace Processor
         }
 
         int process(
-                const unsigned theId, 
+                const unsigned long long theId, 
                 void (*theFunc)());
 
         template<typename ParamType1>
         int process(
-                const unsigned theId, 
+                const unsigned long long theId, 
                 void (*theFunc)(ParamType1),
                 ParamType1 theParam);
         
         template<typename ClassType>
         int process(
-                const unsigned theId, 
+                const unsigned long long theId, 
                 void (ClassType::*theFunc)(),
                 ClassType*const theObj);
 
         template<typename ClassType>
         int process(
-                const unsigned theId, 
+                const unsigned long long theId, 
                 void (ClassType::*theFunc)(),
                 boost::shared_ptr<ClassType> theObj);
 
         template<typename ClassType, typename ParamType1>
         int process(
-                const unsigned theId, 
+                const unsigned long long theId, 
                 void (ClassType::*theFunc)(ParamType1),
                 ClassType* theObj,
                 ParamType1 theParam);
@@ -85,7 +85,7 @@ namespace Processor
                  typename ParamType1,
                  typename ParamType2>
         int process(
-                const unsigned theId, 
+                const unsigned long long theId, 
                 void (ClassType::*theFunc)(ParamType1, ParamType2),
                 ClassType* theObj,
                 ParamType1 theParam1,
@@ -105,7 +105,7 @@ namespace Processor
 
 	template<typename ParamType1>
 	int BoostProcessor::process(
-			const unsigned theId, 
+			const unsigned long long theId, 
 			void (*theFunc)(ParamType1),
 			ParamType1 theParam)
 	{
@@ -121,7 +121,7 @@ namespace Processor
 
     template<typename ClassType>
     int BoostProcessor::process(
-            const unsigned theId, 
+            const unsigned long long theId, 
             void (ClassType::*theFunc)(),
             ClassType*const theObj)
     {
@@ -137,7 +137,7 @@ namespace Processor
 
     template<typename ClassType>
     int BoostProcessor::process(
-            const unsigned theId, 
+            const unsigned long long theId, 
             void (ClassType::*theFunc)(),
             boost::shared_ptr<ClassType> theObj)
     {
@@ -153,7 +153,7 @@ namespace Processor
 
     template<typename ClassType, typename ParamType1>
     int BoostProcessor::process(
-            const unsigned theId, 
+            const unsigned long long theId, 
             void (ClassType::*theFunc)(ParamType1),
             ClassType* theObj,
             ParamType1 theParam)
@@ -172,7 +172,7 @@ namespace Processor
              typename ParamType1,
              typename ParamType2>
     int BoostProcessor::process(
-            const unsigned theId, 
+            const unsigned long long theId, 
             void (ClassType::*theFunc)(ParamType1, ParamType2),
             ClassType* theObj,
             ParamType1 theParam1,
